+++ res/css/views/toasts/_IncomingCallToast.scss
@@ -1,0 +1,100 @@
+/*
+Copyright 2020 The Matrix.org Foundation C.I.C.
+Copyright 2021 Šimon Brandner <simon.bra.ag@gmail.com>
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+.mx_IncomingCallToast {
+    // mx_Toast overrides
+    padding: 8px !important;
+    display: unset !important;
+    top: 8px !important;
+    border-radius: 8px;
+
+    min-width: 250px;
+    box-shadow: 0px 14px 24px rgba(0, 0, 0, 0.08);
+    background-color: $voipcall-plinth-color; // To match mx_Toast
+
+    pointer-events: initial; // restore pointer events so the user can accept/decline
+    cursor: pointer;
+
+    .mx_IncomingCallToast_CallerInfo {
+        display: flex;
+        direction: row;
+
+        img, .mx_BaseAvatar_initial {
+            margin: 8px;
+        }
+
+        > div {
+            display: flex;
+            flex-direction: column;
+
+            justify-content: center;
+        }
+
+        h1, p {
+            margin: 0px;
+            padding: 0px;
+            font-size: $font-14px;
+            line-height: $font-16px;
+        }
+
+        h1 {
+            font-weight: bold;
+        }
+    }
+
+    .mx_IncomingCallToast_buttons {
+        padding: 8px;
+        display: flex;
+        flex-direction: row;
+
+        > .mx_IncomingCallToast_spacer {
+            width: 8px;
+        }
+
+        > * {
+            flex-shrink: 0;
+            flex-grow: 1;
+            margin-right: 0;
+            font-size: $font-15px;
+            line-height: $font-24px;
+        }
+    }
+
+    .mx_IncomingCallToast_iconButton {
+        position: absolute;
+        right: 8px;
+
+        &::before {
+            content: '';
+
+            height: 20px;
+            width: 20px;
+            background-color: $icon-button-color;
+            mask-repeat: no-repeat;
+            mask-size: contain;
+            mask-position: center;
+        }
+    }
+
+    .mx_IncomingCallToast_silence::before {
+        mask-image: url('$(res)/img/voip/silence.svg');
+    }
+
+    .mx_IncomingCallToast_unSilence::before {
+        mask-image: url('$(res)/img/voip/un-silence.svg');
+    }
+}
